// src/firmachain/firmachain.firmachain.nft/module/rest.d.ts
export interface NftNftItem {
    owner?: string;
    /** @format uint64 */
    id?: string;
    tokenURI?: string;
}
export interface NftQueryAllNftItemResponse {
    NftItem?: NftNftItem[];
    /**
     * PageResponse is to be embedded in gRPC response messages where the
     * corresponding request message has used PageRequest.
     *
     *  message SomeResponse {
     *          repeated Bar results = 1;
     *          PageResponse page = 2;
     *  }
     */
    pagination?: V1Beta1PageResponse;
}
export interface NftQueryBalanceOfResponse {
    /** @format uint64 */
    total?: string;
}
export interface NftQueryGetNftItemResponse {
    NftItem?: NftNftItem;
}
export interface NftQueryTokenOfOwnerByIndexResponse {
    /** @format uint64 */
    tokenId?: string;
}
export interface ProtobufAny {
    typeUrl?: string;
    /** @format byte */
    value?: string;
}
export interface RpcStatus {
    /** @format int32 */
    code?: number;
    message?: string;
    details?: ProtobufAny[];
}
// ...
export interface V1Beta1PageResponse {
    /** @format byte */
    nextKey?: string;
    /** @format uint64 */
    total?: string;
}
export declare type QueryParamsType = Record<string | number, any>;
export declare type ResponseFormat = keyof Omit<Body, "body" | "bodyUsed">;
export interface FullRequestParams extends Omit<RequestInit, "body"> {
    /** set parameter to `true` for call `securityWorker` for this request */
    secure?: boolean;
    /** request path */
    path: string;
    /** content type of request body */
    type?: ContentType;
    /** query params */
    query?: QueryParamsType;
    /** format of response (i.e. response.json() -> format: "json") */
    format?: keyof Omit<Body, "body" | "bodyUsed">;
    /** request body */
    body?: unknown;
    /** base url */
    baseUrl?: string;
    /** request cancellation token */
    cancelToken?: CancelToken;
}
export declare type RequestParams = Omit<FullRequestParams, "body" | "method" | "query" | "path">;
export interface ApiConfig<SecurityDataType = unknown> {
    baseUrl?: string;
    baseApiParams?: Omit<RequestParams, "baseUrl" | "cancelToken" | "signal">;
    securityWorker?: (securityData: SecurityDataType) => RequestParams | void;
}
export interface HttpResponse<D extends unknown, E extends unknown = unknown> extends Response {
    data: D;
    error: E;
}
declare type CancelToken = Symbol | string | number;
export declare enum ContentType {
    Json = "application/json",
    FormData = "multipart/form-data",
    UrlEncoded = "application/x-www-form-urlencoded"
}
export declare class HttpClient<SecurityDataType = unknown> {
    baseUrl: string;
    private securityData;
    private securityWorker;
    private abortControllers;
    private baseApiParams;
    constructor(apiConfig?: ApiConfig<SecurityDataType>);
    setSecurityData: (data: SecurityDataType) => void;
    private addQueryParam;
    protected toQueryString(rawQuery?: QueryParamsType): string;
    protected addQueryParams(rawQuery?: QueryParamsType): string;
    private contentFormatters;
    private mergeRequestParams;
    private createAbortSignal;
    abortRequest: (cancelToken: CancelToken) => void;
    request: <T = any, E = any>({ body, secure, path, type, query, format, baseUrl, cancelToken, ...params }: FullRequestParams) => Promise<HttpResponse<T, E>>;
}
/**
 * @title nft/genesis.proto
 * @version version not set
 */
export declare class Api<SecurityDataType extends unknown> extends HttpClient<SecurityDataType> {
    /**
     * No description
     *
     * @tags Query
     * @name QueryBalanceOf
     * @summary Queries a list of balanceOf items.
     * @request GET:/firmachain/firmachain/nft/balanceOf
     */
    queryBalanceOf: (query?: {
        ownerAddress?: string;
    }, params?: RequestParams) => Promise<HttpResponse<NftQueryBalanceOfResponse, RpcStatus>>;
    /**
     * No description
     *
     * @tags Query
     * @name QueryNftItemAll
     * @summary Queries a list of nftItem items.
     * @request GET:/firmachain/firmachain/nft/nftItem
     */
    queryNftItemAll: (query?: {
        "pagination.key"?: string;
        "pagination.offset"?: string;
        "pagination.limit"?: string;
        "pagination.countTotal"?: boolean;
    }, params?: RequestParams) => Promise<HttpResponse<NftQueryAllNftItemResponse, RpcStatus>>;
    /**
     * No description
     *
     * @tags Query
     * @name QueryNftItem
     * @summary Queries a nftItem by id.
     * @request GET:/firmachain/firmachain/nft/nftItem/{id}
     */
    queryNftItem: (id: string, params?: RequestParams) => Promise<HttpResponse<NftQueryGetNftItemResponse, RpcStatus>>;
    /**
     * No description
     *
     * @tags Query
     * @name QueryTokenOfOwnerByIndex
     * @summary Queries a list of tokenOfOwnerByIndex items.
     * @request GET:/firmachain/firmachain/nft/tokenOfOwnerByIndex
     */
    queryTokenOfOwnerByIndex: (query?: {
        ownerAddress?: string;
        index?: string;
    }, params?: RequestParams) => Promise<HttpResponse<NftQueryTokenOfOwnerByIndexResponse, RpcStatus>>;
}